Download data from NCBI and set up SQLite database

Convenience function to do all necessary preparations downloading names, nodes and accession2taxid data from NCBI and preprocessing into a SQLite database for downstream use.

prepareDatabase(sqlFile = "nameNode.sqlite", tmpDir = ".",
  vocal = TRUE, ...)

Arguments

sqlFile
character string giving the file location to store the SQLite database

tmpDir
location for storing the downloaded files from NCBI. (Note that it may be useful to store these somewhere convenient to avoid redownloading)

vocal
if TRUE output messages describing progress

...
additional arguments to getNamesAndNodes, getAccession2taxid or read.accession2taxid

Value

a vector of character string giving the path to the SQLite file

See also

Examples

# NOT RUN {
  if(readline(
    "This will download a lot data and take a while to process.
     Make sure you have space and bandwidth. Type y to continue: "
  )!='y')
    stop('This is a stop to make sure no one downloads a bunch of data unintentionally')

  prepareDatabase()
# }