Read NCBI names file

Take an NCBI names file, keep only scientific names and convert it to a SQLite table

read.names.sql(nameFile, sqlFile = "nameNode.sqlite", overwrite = FALSE)

Arguments

nameFile
string giving the path to an NCBI name file to read from (both gzipped or uncompressed files are ok)

sqlFile
a string giving the path where the output sqlite file should be saved

overwrite
If TRUE, delete names table in database if present and regenerate

Value

invisibly returns a string with path to sqlfile

References

ftp://ftp.ncbi.nih.gov/pub/taxonomy/

See also

Examples

namesText<-c(
  "1\t|\tall\t|\t\t|\tsynonym\t|",
  "1\t|\troot\t|\t\t|\tscientific name\t|",
  "2\t|\tBacteria\t|\tBacteria <prokaryotes>\t|\tscientific name\t|",
  "2\t|\tMonera\t|\tMonera <Bacteria>\t|\tin-part\t|",
  "2\t|\tProcaryotae\t|\tProcaryotae <Bacteria>\t|\tin-part\t|"
)
tmpFile<-tempfile()
writeLines(namesText,tmpFile)
read.names.sql(tmpFile)

#> nameNode.sqlite already contains table names. Delete file or set overwrite=TRUE to reload