Read NCBI nodes file

Take an NCBI nodes file and convert it to a data.table

read.nodes2(nodeFile, sqlFile = "nameNode.sqlite")

Arguments

nodeFile
string giving the path to an NCBI node file to read from (both gzipped or uncompressed files are ok)

sqlFile
a string giving the path where the output sqlite file should be saved

Value

a data.table with columns id, parent and rank with a key on id

References

ftp://ftp.ncbi.nih.gov/pub/taxonomy/

See also

read.names

Examples

nodes<-c(
 "1\t|\t1\t|\tno rank\t|\t\t|\t8\t|\t0\t|\t1\t|\t0\t|\t0\t|\t0\t|\t0\t|\t0\t|\t\t|",
 "2\t|\t131567\t|\tsuperkingdom\t|\t\t|\t0\t|\t0\t|\t11\t|\t0\t|\t0\t|\t0\t|\t0\t|\t0\t|\t\t|",
 "6\t|\t335928\t|\tgenus\t|\t\t|\t0\t|\t1\t|\t11\t|\t1\t|\t0\t|\t1\t|\t0\t|\t0\t|\t\t|",
 "7\t|\t6\t|\tspecies\t|\tAC\t|\t0\t|\t1\t|\t11\t|\t1\t|\t0\t|\t1\t|\t1\t|\t0\t|\t\t|",
 "9\t|\t32199\t|\tspecies\t|\tBA\t|\t0\t|\t1\t|\t11\t|\t1\t|\t0\t|\t1\t|\t1\t|\t0\t|\t\t|"
)
tmp<-tempfile()
read.nodes2(textConnection(nodes),tmp)